backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Config
pydantic-settings==2.4.0

# Supabase
supabase==2.9.0

# Redis
redis[asyncio]==5.0.8

# Kafka
aiokafka==0.11.0

# JWT
python-jose[cryptography]==3.3.0

# HTTP client
httpx==0.27.0

# Razorpay
razorpay==1.4.2

# Utilities
python-multipart==0.0.9  # form data